+++ src/index.jsx
@@ -4,13 +4,11 @@
 import React from "react";
 import {render} from "react-dom";
 import {Provider} from "react-redux";
-import {browserHistory} from "react-router";
 import App from "./components/App";
 
 import configureStore from "./store/configureStore";
 
 import "./styles/styles.scss";
-import {syncHistoryWithStore} from "react-router-redux";
 
 require("./public/favicon.ico");
 require("./images/card_bg.jpg");
@@ -20,11 +18,8 @@ require("./images/map-marker.png");
 
 const store = configureStore();
 
-// Create an enhanced history that syncs navigation events with the store
-syncHistoryWithStore(browserHistory, store);
-
 render(
 	<Provider store={store}>
-		<App />
+			<App />
 	</Provider>, document.getElementById("app")
 );