+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changed
+- Upgraded react-router and connected react-router packages. Fixed some vulnerabilities [CCROP-379](https://opensource.ncsa.illinois.edu/jira/browse/CCROP-379)
+
 ## [1.4.0] - 2021-10-26
 
 ### Added