+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [1.5.0] - 2021-12-16
 
 ### Added
 - Ability to update tile drainage from "Field Profile" tab. [CCROP-292](https://opensource.ncsa.illinois.edu/jira/browse/CCROP-292)